#step1 [refactor] parse 메소드 분리
add 메소드 내부에서 stream 및 mapToInt를 활용하여 string -> int로 변환하는 부분을 parse 메소드로 분리
